pyproject.toml
[project]
name = "spam-filter"
version = "2.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.13"
dependencies = [
    "beautifulsoup4==4.12.2",
    "combo==0.1.3",
    "dill==0.3.7",
    "nltk==3.8.1",
    "python-telegram-bot>=22.5",
    "scikit-learn==1.3.0",
    "scipy==1.11.4",
    "urlextract==1.8.0",
]
